:art: shared/ui 파일 분리
